pyproject.toml
[tool.poetry]
name = "boiler"
version = "0.1.0"
description = ""
authors = ["alish-xeno <charitra.prajapati@featherwebs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2 = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
flask-restful = "^0.3.10"
marshmallow = "^3.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
